+++ src/components/styles/StyledButton.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const StyledButton = styled.div`
+  display: inline-block;
+  cursor: pointer;
+`;
+
+export default StyledButton;